requirements.txt
Flask
openai-whisper
tqdm
gunicorn
